refactor: remove unused inventory model and clean up notification and input services
