+++ src/routes/baseItems.route.js
@@ -1,9 +1,9 @@
 const express = require("express");
-const AuthMiddleware = require("../../middlewares/auth.middleware");
-const { USER_ROLES } = require("../../configs/user.config");
-const checkRoles = require("../../middlewares/role.middleware");
-const { catchAsyncHandle } = require("../../middlewares/error.middleware");
-const baseItemsController = require("../../controllers/baseItems.controller");
+const AuthMiddleware = require("../middlewares/auth.middleware");
+const { USER_ROLES } = require("../configs/user.config");
+const checkRoles = require("../middlewares/role.middleware");
+const { catchAsyncHandle } = require("../middlewares/error.middleware");
+const baseItemsController = require("../controllers/baseItems.controller");
 
 const router = express.Router();
 router.use(
