+++ src/routes/index.js
@@ -23,260 +23,6 @@ const warehouseStorageModel = require("../models/warehouseStorage.model");
 // const systemModel = require("../models/system.model");
 
 router.get("/", async (req, res) => {
-    // await warehouseStorageModel.create({
-    //     warehouseId: new mongoose.Types.ObjectId("67b59863ea76ca0be173fa3b"),
-    //     itemId: new mongoose.Types.ObjectId("67ca7dfc240a67e748ab4c49"),
-    //     batchNumber: "123",
-    //     quantity: 100,
-    // })
-    // const system=await systemModel.create({
-    //     checkExpiredMedicineInterval:'0 0 * * *',
-    // })
-
-    // const admin = await userModel.create({
-    //     fullName: "Admin 001",
-    //     role: "Admin",
-    //     email: "admin001@gmail.com",
-    //     password: await bcrypt.hash("Admin@001", 10)
-    // })
-
-    // const manager = await userModel.create({
-    //     fullName: "Manager 001",
-    //     role: "Manager",
-    //     email: "manager001@gmail.com",
-    //     password: await bcrypt.hash("Manager@001", 10)
-    // })
-
-    // const reportStaff = await userModel.create({
-    //     fullName: "Report Staff 001",
-    //     role: "Report Staff",
-    //     email: "reportStaff001@gmail.com",
-    //     password: await bcrypt.hash("ReportStaff@001", 10)
-    // })
-
-    // const inventoryStaff = await userModel.create({
-    //     fullName: "Inventory Staff 001",
-    //     role: "Inventory Staff",
-    //     email: "inventoryStaff001@gmail.com",
-    //     password: await bcrypt.hash("InventoryStaff@001", 10)
-    // })
-
-    // const supplier = await userModel.create({
-    //     fullName: "Supplier 001",
-    //     role: "Supplier",
-    //     email: "supplier001@gmail.com",
-    //     password: await bcrypt.hash("Supplier@001", 10)
-    // })
-
-    // const customer = await userModel.create({
-    //     fullName: "Customer 001",
-    //     role: "Customer",
-    //     email: "customer001@gmail.com",
-    //     password: await bcrypt.hash("Customer@001", 10)
-    // })
-
-    // const warehouse = await warehouseModel.create({
-    //     name: "Normal Warehouse",
-    //     description: "Normal warehouse for normal medicine and equipment",
-    //     category: "Normal",
-    //     minTemperature: 0,
-    //     maxTemperature: 100,
-    //     status: "Available",
-    // })
-
-    // const diposalWarehouse = await warehouseModel.create({
-    //     name: "Disposal Warehouse",
-    //     description: "Disposal warehouse for expired medicine",
-    //     category: "Disposal",
-    //     status: "Available",
-    //     minTemperature: 0,
-    //     maxTemperature: 100,
-    // })
-
-    // const frozenWarehouse = await warehouseModel.create({
-    //     name: "Cold Warehouse",
-    //     description: "Cold warehouse for cold stored medicine",
-    //     category: "Cold",
-    //     status: "Available",
-    //     minTemperature: -10,
-    //     maxTemperature: 0,
-    // })
-
-    // const baseItem = await baseItemModel.create({
-    //     name: "Paracetamol",
-    //     description: "Paracetamol is a common painkiller used to treat aches and pain. It can also be used to reduce a high temperature. It's available combined with other painkillers and anti-sickness medicines.",
-    //     category: "Medicine",
-    //     genericName: "Paracetamol 500mg Tablets",
-    //     brand: "Panadol",
-    //     countryOfOrigin: "Vietnam",
-    //     indication: "Paracetamol is used to relieve mild to moderate pain from a headache, toothache, cold, flu, joint pain, or periods pain.",
-    //     contraindication: "Paracetamol is contraindicated in patients with a known hypersensitivity to paracetamol.",
-    //     sideEffect: "Paracetamol is usually well tolerated at recommended doses. The most common side effects are usually mild. These include an allergic reaction, which may cause a rash and sometimes swelling.",
-    //     storageType: "Normal",
-    // });
-
-    // const item = await itemModel.create({
-    //     baseItemId: baseItem._id,
-    //     code: "PCT-001",
-    //     status: "Available",
-    //     manufactureDate: new Date() - 1000 * 60 * 60 * 24 * 365,
-    //     expiredDate: new Date(),
-    //     unit: "Tablet",
-    // })
-
-    // const input = await inputModel.create({
-    //     description: "Input from Tiki",
-    //     status: "Pending",
-    //     batchNumber:
-    //         new Date().getTime().toString() +
-    //         "-INP",
-    //     warehouseId: warehouse._id,
-    //     managerId: manager._id,
-    //     supplierId: supplier._id,
-    //     inventoryStaffId: inventoryStaff._id,
-    // })
-
-    // const inputDetail = await inputDetailModel.create({
-    //     inputId: input._id,
-    //     itemId: item._id,
-    //     quantity: 100,
-    //     inputPrice: 10000,
-    //     status: "Pending",
-    // })
-
-    // const inventory = await inventoryModel.create({
-    //     warehouseId: warehouse._id,
-    //     itemId: item._id,
-    //     quantity: 100,
-    //     batchNumber:
-    //         new Date().getTime().toString() +
-    //         '-INV-' +
-    //         item.code,
-    // })
-
-    // await stockTransactionModel.create({
-    //     warehouseId: warehouse._id,
-    //     itemId: item._id,
-    //     transactionType: "Input",
-    //     quantity: 100,
-    //     reason: "Import from Tiki",
-    // })
-
-    // const output = await outputModel.create({
-    //     description: "Output to Customer",
-    //     status: "Pending",
-    //     batchNumber:
-    //         new Date().getTime().toString() +
-    //         '-OUP',
-    //     warehouseId: warehouse._id,
-    //     managerId: manager._id,
-    //     customerId: customer._id,
-    //     inventoryStaffId: inventoryStaff._id,
-    // })
-
-    // const outputDetail = await outputDetailModel.create({
-    //     outputId: output._id,
-    //     itemId: item._id,
-    //     quantity: 20,
-    //     outputPrice: 12000,
-    // })
-
-    // await stockTransactionModel.create({
-    //     warehouseId: warehouse._id,
-    //     itemId: item._id,
-    //     transactionType: "Output",
-    //     quantity: 20,
-    //     description: "Export to Customer",
-    // })
-
-    // await inventoryModel.updateOne({ _id: inventory._id }, { quantity: 80 })
-
-    // const stockCheck = await stockCheckModel.create({
-    //     warehouseId: warehouse._id,
-    //     status: "Pending",
-    //     managerId: manager._id,
-    //     inventoryStaffId: inventoryStaff._id,
-    //     description: "Stock Check",
-    // })
-
-    // const stockCheckDetail = await stockCheckDetailModel.create({
-    //     stockCheckId: stockCheck._id,
-    //     itemId: item._id,
-    //     systemQuantity: 80,
-    //     actualQuantity: 80,
-    //     difference: 0,
-    // })
-
-    // const warehouseTransaction = await warehouseTransactionModel.create({
-    //     fromWarehouseId: warehouse._id,
-    //     toWarehouseId: diposalWarehouse._id,
-    //     managerId: manager._id,
-    //     inventoryStaffId: inventoryStaff._id,
-    //     description: "Transfer to Disposal Warehouse",
-    //     status: "Pending",
-    // })
-
-    // const warehouseTransactionDetail = await warehouseTransactionDetailModel.create({
-    //     warehouseTransactionId: warehouseTransaction._id,
-    //     itemId: item._id,
-    //     quantity: 20,
-    // })
-
-    // await inventoryModel.updateOne({
-    //     _id: inventory._id
-    // }, { quantity: 60 })
-    // await stockTransactionModel.create({
-    //     warehouseId: warehouse._id,
-    //     itemId: item._id,
-    //     transactionType: "Output",
-    //     quantity: 20,
-    //     description: "Transfer to Disposal Warehouse",
-    // })
-
-    // const diposalInventory = await inventoryModel.create({
-    //     warehouseId: diposalWarehouse._id,
-    //     itemId: item._id,
-    //     quantity: 20
-    // })
-
-    // await stockTransactionModel.create({
-    //     warehouseId: diposalWarehouse._id,
-    //     itemId: item._id,
-    //     transactionType: "Input",
-    //     quantity: 20,
-    //     description: "Transfer from Medical Warehouse",
-    // })
-
-    // const warehouseCheck = await warehouseCheckModel.create({
-    //     warehouseId: frozenWarehouse._id,
-    //     status: "Pending",
-    //     managerId: manager._id,
-    //     inventoryStaffId: inventoryStaff._id,
-    //     description: "Check Frozen Warehouse Temperature",
-    // })
-
-    // const warehouseCheckDetail = await warehouseCheckDetailModel.create({
-    //     warehouseCheckId: warehouseCheck._id,
-    //     description: "Check Frozen Warehouse Temperature",
-    //     temperature: -5,
-    //     thresholdLevel: "Normal",
-    //     condition: "Good",
-    //     status: "Pending"
-    // })
-
-    // const report = await reportModel.create({
-    //     managerId: manager._id,
-    //     reportStaffId: reportStaff._id,
-    //     title: "Stock Report",
-    //     content: "Stock Report Content",
-    // })
-
-    // const reportDetail = await reportDetailModel.create({
-    //     reportId: report._id,
-    //     type: "Inventory",
-    //     detail: "Stock Report Detail",
-    // })
-
     res.status(200).send(`
 <!DOCTYPE html>
 <html lang="en">
@@ -411,13 +157,13 @@ router.get("/", async (req, res) => {
     res.end();
 });
 
-router.use("/auth", require("./auth"));
-router.use("/users", require("./users"));
-router.use("/items", require("./items"));
-router.use("/baseitems", require("./baseItems"));
-router.use("/system", require("./system"));
-router.use("/outputs", require("./outputs"));
-router.use("/warehouses", require("./warehouses"));
-router.use("/inputs", require("./inputs"));
+router.use("/auth", require("./auth.route"));
+router.use("/users", require("./users.route"));
+router.use("/items", require("./items.route"));
+router.use("/baseitems", require("./baseItems.route"));
+router.use("/system", require("./system.route"));
+router.use("/outputs", require("./outputs.route"));
+router.use("/warehouses", require("./warehouses.route"));
+router.use("/inputs", require("./inputs.route"));
 
 module.exports = router;
