+++ src/routes/items.route.js
@@ -1,9 +1,9 @@
 const express = require("express");
-const checkRoles = require("../../middlewares/role.middleware");
-const AuthMiddleware = require("../../middlewares/auth.middleware");
-const { catchAsyncHandle } = require("../../middlewares/error.middleware");
-const { USER_ROLES } = require("../../configs/user.config");
-const itemController = require("../../controllers/item.controller");
+const checkRoles = require("../middlewares/role.middleware");
+const AuthMiddleware = require("../middlewares/auth.middleware");
+const { catchAsyncHandle } = require("../middlewares/error.middleware");
+const { USER_ROLES } = require("../configs/user.config");
+const itemController = require("../controllers/item.controller");
 
 const router = express.Router();
