+++ src/routes/warehouses.route.js
@@ -1,9 +1,9 @@
 const express = require("express");
-const { catchAsyncHandle } = require("../../middlewares/error.middleware");
-const warehouseController = require("../../controllers/warehouse.controller");
-const AuthMiddleware = require("../../middlewares/auth.middleware");
-const checkRoles = require("../../middlewares/role.middleware");
-const { USER_ROLES } = require("../../configs/user.config");
+const { catchAsyncHandle } = require("../middlewares/error.middleware");
+const warehouseController = require("../controllers/warehouse.controller");
+const AuthMiddleware = require("../middlewares/auth.middleware");
+const checkRoles = require("../middlewares/role.middleware");
+const { USER_ROLES } = require("../configs/user.config");
 
 const router = express.Router();
